Windows portability: fix MinGW build errors
- fs::path::c_str() returns wchar_t* on Windows — use .string().c_str()
- Guard POSIX socket/glob code with #ifndef _WIN32 (dbg_io, dbg_client)
- Guard debug server in main.cpp with #ifndef _WIN32
- mkdir on Windows takes 1 arg (no permissions)
- Include GL/glext.h on non-Apple for GL_BGRA etc.
- Replace %zu with %lu + cast for MinGW printf compatibility
- Add __USE_MINGW_ANSI_STDIO=1 for MinGW builds

+++ src/devices/slip.h
@@ -46,7 +46,8 @@ inline void encode(const uint8_t *pkt, size_t len, std::vector<uint8_t> &out)
 		}
 	}
 	out.push_back(END);
-	SLP_LOG("encode: %zu bytes -> %zu framed", len, out.size());
+	SLP_LOG("encode: %lu bytes -> %lu framed", static_cast<unsigned long>(len),
+			static_cast<unsigned long>(out.size()));
 }
 
 /*
@@ -63,7 +64,8 @@ public:
 		{
 			if (!accum_.empty())
 			{
-				SLP_LOG("decode: complete packet %zu bytes", accum_.size());
+				SLP_LOG("decode: complete packet %lu bytes",
+						static_cast<unsigned long>(accum_.size()));
 				return true; /* packet complete */
 			}
 			return false; /* empty frame (inter-packet END) — ignore */
